Move indexer queue functionality from client to service
